alter: fix an assert failure when altering a collation
Forbid update in _collation system space. It leads to an assertion
failure in collation cache, since it has no support for replace
operation.
Refactor collation deletion and insertion in alter.cc.

Besides, collation update potentially changes order of data
in the index and did not trigger rebuild of all indexes.