sql: IS is only applicable when dealing with NULL
According to ANSI Standard IS/IS NOT can be used to determine
if values is null. At the same time in SQLite3 IS/IS NOT have an
additional function - it can be used to check equality of two
values. This feature isn't common for different versions of SQL
(only in PL/SQL right operand can be NONE, TRUE of FALSE) This
patch removes described function.