sql: VDBE tests for trigger existence
Trigger presence in system should be tested on each VDBE
execution attempt, not on Parser iteration as system state
could be changed between opcode run.
With this patch, such checks are the part of VDBE program
looking to the _trigger space and raise error message if
tuple with specified key already exists.

Part of #3435, #3273