Detect when instance is run or restarted by tarantoolctl.
There are some hacks to know the instance was run by tarantoolctl,
none of them are too reliable, though. This patch introduces 2
environment variables set by tarantoolctl for the instance to
know when it's being run or restarted.

Closes: #3215

@TarantoolBot document
Title: tarantoolctl: document setting environment variables
tarantoolctl sets the `TARANTOOLCTL` environment variable when starting
an instance, and sets the `TARANTOOL_RESTARTED' environment variable
when restarting.