vinyl: pin last statement returned by write iterator explicitly
Currently, the last statement returned by the write iterator is
referenced indirectly, via a read view. This works, because the write
iterator can only return a statement if it corresponds to a certain read
view.  However, in the scope of #2129, the write iterator will also have
to keep statements for which a deferred DELETE hasn't been generated
yet, even if no read view needs it. So let's make the write iterator
reference the last returned statement explicitly, i.e. via a dedicated
member of the write_iterator struct.

Needed for #2129