Use ON_CONFLICT_ACTION_DEFAULT for non-nullable key parts
Currently, we set nullable_action to ON_CONFLICT_ACTION_ABORT for key
parts that have is_nullable set to false, although for field definitions
we use ON_CONFLICT_ACTION_DEFAULT in this case. This is confusing.

Actually, we are supposed to set ON_CONFLICT_ACTION_DEFAULT if no action
was explicitly specified and the index/column is not nullable. So let's
use ON_CONFLICT_ACTION_DEFAULT for key parts as well.

Suggested by @kyukhin.