vinyl: abort rw transactions before DDL
When building a new index or checking a space format, we propagate
changes done to the space during the procedure with the aid of an
on_replace trigger. The problem is there may be transactions with a
non-empty write set when we install the trigger. Changes done by
those transactions will not be seen by the trigger and so they won't
make it to the new index, resulting in an inconsistency between the
primary and secondary indexes. To fix this issue, let's abort all rw
transactions affecting the space after setting the trigger. Note,
transactions that reached WAL can't be aborted so we wait for them to
complete.

Closes #3458