test: fix box/net_box hang when run after box/access_escalation
box/access_escalation cancels a fiber that is waiting for a call over
iproto to complete, but the fiber that is executing the call keeps
running. As a result, if box/net_box is run after this test, it will
hang, because it expects that the fiber pool is idle when it starts.
Fix this by making the call function return in box/access_escalation.