test: fix sporadic vinyl/ddl hang
There's a case in the test that expects that when we dump a new run,
vinyl will trigger compaction. Typically, it happens, but sometimes,
presumably due to compression, the new run happens to be a bit smaller
than the old one so that compaction isn't triggered, and the test hangs.
Fix this by forcing compaction with index.compact() knob.