test: fix engine/ddl occasional failure
Sometimes the test failed with output similar to the one below:

    [001] engine/ddl.test.lua	memtx           [ fail ]
    [001]
    [001] Test failed! Result content mismatch:
    [001] --- engine/ddl.result	Mon Aug 27 09:35:19 2018
    [001] +++ engine/ddl.reject	Mon Aug 27 11:12:47 2018
    [001] @@ -1932,7 +1932,7 @@
    [001]  ...
    [001]  s.index.pk:select()
    [001]  ---
    [001] -- - [1, 1, 11]
    [001] +- - [1, 1, 8]
    [001]  ...
    [001]  s.index.sk:select()
    [001]  ---

This happened due to a race condition in a test case added for
issue #3578.
To fix it we need to move c:get() above s.index.pk:select() to
make sure we actually wait for the fiber function to complete
before checking results.

Follow-up #3578.