swim: introduce failure detection component
Failure detection components allows to find which members are
already dead. It works via pings attached to each round message.
A member, received ping, should respond with an ack. If it is not
done within a timeout, then the ping is considered unacknowledged.
When a number of pings grows to a threshold, the member is
declared DEAD. After more unacked pings it is dropped from the
member table. Unless the GC was turned off.

Part of #3234