box: cleanup region after box.snapshot()
Before this patch region wasn't truncated after box.snapshot() so some
tests failed due to "memory leak". To fix that, let's truncate region in
vinyl and use malloc() in memtx (we can't cleanup region there as
objects allocated when checkpoint is started are used until checkpoint
is completed). The latter isn't a big deal, because box.snapshot() isn't
a hot path and we do other allocations with malloc() anyway.

Closes #3732