json: introduce json_path_cmp, json_path_validate
Introduced json_path_validate routine to ensure user-defined
JSON path is valid. This will be required to raise an error if
an incorrect user-defined jason-path is detected.

Introduced json_path_cmp routine to compare JSON paths that may
have different representation.
Note that:
 - in case of paths that have same token-sequence prefix,
   the path having more tokens is assumed to be greater
 - both paths to compare must be valid

@locker: move json_path_validate to the object file (no need to
pollute the header with this cold routine).

Needed for #1012