sql: disallow to rename space if it is referenced by view
Before this patch it was allowed to rename space which is referenced by
a view. In turn, view contains SELECT statement in a raw form (i.e. as a
string) and it is not modified during renaming routine. Hence, after
renaming space still has referencing counter > 0, but no usage of view
is allowed (since execution of SELECT results in "Space does not
exist"). To avoid such situations, lets ban renaming space if its view
reference counter > 0.

Note that RENAME is ANSI extension, so different DBs behave in this case
in different ways - some of them allow to rename tables referenced by a
view (PostgreSQL), others - don't (Oracle).

Closes #3746