vinyl: fix tuple leak in vy_tuple_new
tuple_format_vtab::tuple_new() callback is supposed to return a tuple
with 0 references, but vinyl implementation returns a tuple with the
reference counter set to 1, effectively leaking the new tuple.

How to reproduce: get a tuple from a vinyl space and update it with
box.tuple.update().

The bug was introduced by commit a377332a7314 ("tuple: fix lost format
on update and upsert operations").